adding javapoet generator
