+++ client/plugin/src/main/java/io/openlineage/client/OpenlineageClientPlugin.java
@@ -1,0 +1,19 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package io.openlineage.client;
+
+import org.gradle.api.Project;
+import org.gradle.api.Plugin;
+
+/**
+ * A simple 'hello world' plugin.
+ */
+public class OpenlineageClientPlugin implements Plugin<Project> {
+    public void apply(Project project) {
+        // Register a task
+        project.getTasks().register("greeting", task -> {
+            task.doLast(s -> System.out.println("Hello from plugin 'io.openlineage.client.greeting'"));
+        });
+    }
+}
